fix: moved prometheus metrics for a specific paste
